Final python commit

+++ python/azurefunc/ParseLog/__init__.py
@@ -6,6 +6,29 @@ import azure.functions as func
 
 import pymongo
 
+def find_bldg_from_id(ap):
+    bldg = re.search("(^[A-Za-z]+)", ap)[0]
+
+    uri = "mongodb://localhost:27017/ninerfi"# os.environ["COSMOS_CONNECTION_STRING"]
+    client = pymongo.MongoClient(uri)
+
+    db = client.ninerfi
+    collection = db.access_points
+
+    result = collection.find_one({"prefix": bldg})
+    if (result):
+        return result['_id']
+    
+    if (collection.count_documents({"prefix": ""}) > 0):
+        result = collection.find({"prefix": ""})
+        for doc in result:
+            if (doc['building'].startswith(bldg)):
+                collection.update_one({"_id": doc['_id']}, {"$set": {"prefix": bldg}})
+                return doc['_id']
+    
+    logging.info('MANUAL --- ' + bldg)
+    return None
+
 
 def main(req: func.HttpRequest) -> func.HttpResponse:
     logging.info('Python HTTP trigger function processed a request.')
@@ -31,7 +54,7 @@ def main(req: func.HttpRequest) -> func.HttpResponse:
             exited_ap = re.search("<\s(?:[0-9]{1,3}\.){3}[0-9]{1,3}>\s\s(Deauth to sta): (.{17}).+?AP.+?(?:EXT-)?((?:[A-Za-z]){3,}[^\s]+)", log['log'])
 
             if (entered_ap):
-                uri = os.environ["COSMOS_CONNECTION_STRING"]
+                uri = "mongodb://localhost:27017/ninerfi"#os.environ["COSMOS_CONNECTION_STRING"]
                 client = pymongo.MongoClient(uri)
 
                 db = client.ninerfi
@@ -40,17 +63,17 @@ def main(req: func.HttpRequest) -> func.HttpResponse:
                 mac_address = entered_ap[2]
                 ap = entered_ap[3]
 
-                collection.update_one({"mac_address": mac_address}, {"$set": {
-                    "access_point": {
-                        "id": ap,
-                        "building": re.search("(^[A-Za-z]+)", ap)[0]
-                    }
-                }}, upsert=True)
+                apId = find_bldg_from_id(ap)
 
-                logging.info(f'Access Point {ap} count updated.')
+                if (apId):
+                    collection.update_one({"mac_address": mac_address}, {"$set": {
+                        "access_point": apId
+                    }}, upsert=True)
+
+                    logging.info(f'Access Point {ap} count updated.')
             
             elif (exited_ap):
-                uri = os.environ["COSMOS_CONNECTION_STRING"]
+                uri = "mongodb://localhost:27017/ninerfi"#os.environ["COSMOS_CONNECTION_STRING"]
                 client = pymongo.MongoClient(uri)
 
                 db = client.ninerfi
@@ -59,9 +82,12 @@ def main(req: func.HttpRequest) -> func.HttpResponse:
                 mac_address = exited_ap[2]
                 ap = exited_ap[3]
 
-                result = collection.delete_one({"mac_address": mac_address})
+                apId = find_bldg_from_id(ap)
 
-                if (result.deleted_count > 0): logging.info(f'Access Point {ap} count updated.')
+                if (apId):
+                    result = collection.delete_one({"mac_address": mac_address})
+
+                    if (result.deleted_count > 0): logging.info(f'Access Point {ap} count updated.')
             
             else:
                 pass
